requirements.txt
flask
requests
pyOpenSSL
waitress
twilio
python-telegram-bot